Revert "Feature/reorganize screens"
