+++ index.html
@@ -3,9 +3,7 @@
     <title>Project Swimmy Mock-Up</title>
     <link rel="stylesheet" type="text/css" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
     <link rel="stylesheet" type="text/css" href="node_modules/font-awesome/css/font-awesome.min.css">
-    <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700" rel="stylesheet">
     <link rel="stylesheet" type="text/css" href="assets/css/styles.css">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
   <body>
     <div id="root"></div>
