+++ package.json
@@ -39,7 +39,6 @@
   },
   "dependencies": {
     "bootstrap": "^3.3.7",
-    "css-loader": "^0.26.1",
     "font-awesome": "^4.7.0",
     "jquery": "^3.1.1",
     "mobx": "^2.2.2",
@@ -47,7 +46,6 @@
     "mobx-react-devtools": "^4.2.0",
     "mobx-router": "0.0.7",
     "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "style-loader": "^0.13.1"
+    "react-dom": "^15.1.0"
   }
 }
