--- src/screens/styles/loginScreen.css
@@ -1,7 +1,0 @@
-.input {
-  composes: input from "../../shared/styles/buttonStyles.css"
-}
-
-.buttons {
-  composes: btn from "../../shared/styles/buttonStyles.css"
-}
