--- src/screens/styles/splashScreenStyles.css
@@ -1,23 +1,0 @@
-.splashScreen {
-  color: black;
-}
-
-.titleText {
-  composes: normalText from "../../shared/styles/fontStyles.css";
-  padding: 18px;
-  text-align: center;
-  margin: 0;
-}
-
-.logoPlaceholder {
-  height: 300px;
-  background-color: #f1f1f1
-}
-
-.buttons {
-  composes: btn from "../../shared/styles/buttonStyles.css"
-}
-
-.buttonContainer {
-  padding: 0 20px;
-}
