--- src/shared/styles/fontStyles.css
@@ -1,40 +1,0 @@
-.h1, .h2, .h3, .h4, .h5, .h6 {
-  padding: 0;
-  margin: 0;
-  line-height: 135%;
-  font-family: 'Montserrat', sans-serif;
-}
-
-.h1 {
-  font-size: 24px;
-  font-weight: 400;
-}
-
-.h2 {
-  font-size: 20px;
-  font-weight: 400;
-}
-
-.h3 {
-  font-size: 18px;
-  font-weight: 400;
-}
-
-.h4 {
-  font-size: 16px;
-  font-weight: 400;
-}
-
-.h5 {
-  font-size: 14px;
-  font-weight: 400;
-}
-
-.h6 {
-  font-size: 12px;
-  font-weight: 400;
-}
-
-.normalText {
-  font-size: 20px;
-}
